backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
httpx>=0.25.0
openenv-core>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
torch>=2.0.0
